fix_(test, mul, div, add, sub): updated bi_print, bi_add, bi_sub, bi_mul, and bi_division to support both uint32_t and uint64_t as the word data type_NOV_19

+++ dtype.h
@@ -9,8 +9,10 @@
 
 typedef int msg;
 typedef uint8_t byte;
-typedef uint16_t half_word;
-typedef uint32_t word;
+// typedef uint16_t half_word;
+// typedef uint32_t word;
+typedef uint32_t half_word;
+typedef uint64_t word;
 
 /**
  * @struct bigint
